Optionally disable managing the GPG public key
Short-circuit checking the keyring for the public key by providing a config option to just opt-out of all the GPG public key management.

If you change this setting, you know you need to bake out the public key in your images/servers and that things may explode on you if you don't.