GPG sign artifact using configured key
GPG defaults to using the first secret key if you don't override it in the (detached) sign command. This may not always be the key you want to sign your pupistry artifact with.

Instead, as part of the --detached-sign command we now pass the `--local-user` flag with the name of the pupistry gpg key from our configuration. This ensures we're signing the artifact with the same key the servers are configured to verify the signature against.